+++ addOns/ascanrulesBeta/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Maintenance Changes.
 - Backup File Disclosure: don't raise issues for non-success codes unless at LOW threshold (Issue 6059).
 - ELMAH Information Leak: don't raise issues unless content looks good unless at LOW threshold (Issue 6076).
+- Session Fixation scan rule fix potential false positive on session cookie HttpOnly, and Secure flags (Issue 6082).
 
 ## [28] - 2020-06-01
 ### Added